+++ app/src/main/java/nodomain/freeyourgadget/gadgetbridge/util/FileUtils.java
@@ -8,6 +8,7 @@ import android.support.annotation.NonNull;
 import android.util.Log;
 
 import java.io.BufferedInputStream;
+import java.io.BufferedReader;
 import java.io.ByteArrayOutputStream;
 import java.io.File;
 import java.io.FileInputStream;
@@ -15,6 +16,7 @@ import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
+import java.io.InputStreamReader;
 import java.nio.channels.FileChannel;
 import java.util.ArrayList;
 import java.util.List;
@@ -62,6 +64,20 @@ public class FileUtils {
         }
         fin.close();
         fout.close();
+    }
+
+    public static String getStringFromFile(File file) throws IOException {
+        FileInputStream fin = new FileInputStream(file);
+
+        BufferedReader reader = new BufferedReader(new InputStreamReader(fin));
+        StringBuilder sb = new StringBuilder();
+        String line;
+        while ((line = reader.readLine()) != null) {
+            sb.append(line).append("\n");
+        }
+        reader.close();
+        fin.close();
+        return sb.toString();
     }
 
     /**